#100, Multistate.js
Here is the javascript code that directly rewrites the elements in the multistate class using the class name as an anchor to access the number box's "disabled" property. Some lines are added for formatting purposes and to prevent too much blank space from accumulating every time the button is clicked. No edits to the index.html were needed.